+++ servant-docs/CHANGELOG.md
@@ -3,6 +3,7 @@ HEAD
 
 * Support for the `HttpVersion`, `IsSecure`, `RemoteHost` and `Vault` combinators
 * Support maximum samples setting with new `DocOptions` type (used by `docsWithOptions` and `docsWith`)
+* Remove redundant second parameter of ToSample
 * Add Generic-based default implementation for `ToSample` class
 * Add more `ToSamples` instances: `Bool`, `Ordering`, tuples (up to 7), `[]`, `Maybe`, `Either`, `Const`, `ZipList` and some monoids
 * Move `toSample` out of `ToSample` class