+++ servant-docs/CHANGELOG.md
@@ -8,6 +8,7 @@ HEAD
 * Add more `ToSamples` instances: `Bool`, `Ordering`, tuples (up to 7), `[]`, `Maybe`, `Either`, `Const`, `ZipList` and some monoids
 * Move `toSample` out of `ToSample` class
 * Add a few helper functions to define `toSamples`
+* Remove matrix params.
 
 0.4
 ---